+++ packages/chip/ChipSet.svelte
@@ -1,0 +1,3 @@
+<div class='gp-chip-set'>
+  <slot></slot>
+</div>