+++ packages/chip/base.js
@@ -1,0 +1,5 @@
+import Chip from "./Chip.svelte";
+import ChipSet from "./ChipSet.svelte";
+
+export default Chip;
+export { Chip, ChipSet };