+++ packages/icons/Cancel.svelte
@@ -1,0 +1,9 @@
+   <!-- This icon was adapted from the Photon Design System (licensed under the MPL v2.0):
+    https://design.firefox.com/photon/
+    https://github.com/FirefoxUX/photon-icons/blob/master/icons/desktop/cancel-16.svg
+   -->
+<script>
+export let size = 24;
+</script>
+
+<svg xmlns="http://www.w3.org/2000/svg" width={size} height={size} viewBox="0 0 16 16"><path fill="currentColor" d="M6.586 8l-2.293 2.293a1 1 0 0 0 1.414 1.414L8 9.414l2.293 2.293a1 1 0 0 0 1.414-1.414L9.414 8l2.293-2.293a1 1 0 1 0-1.414-1.414L8 6.586 5.707 4.293a1 1 0 0 0-1.414 1.414L6.586 8zM8 0a8 8 0 1 1 0 16A8 8 0 0 1 8 0z"></path></svg>